Backend/yolov4/requirements.txt
keras==2.3.1
tensorflow==1.15.4
opencv-python
Pillow
matplotlib
numpy==1.18.5
libSBOL